+++ CHANGELOG.md
@@ -8,7 +8,8 @@ document.
 ### Added
 - New drawImage function in UIRenderer.
 - New width and height members in Texture class.
-- New red frozen vignette ui effect when player is frozen.	
+- New red frozen vignette ui effect when player is frozen.
+- New AudioInstance class to play audio.	
 
 ## [0.3.1] - 2026-09-05